First Tests migrated

+++ packages/stunner-editors/kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-api/src/main/java/org/kie/workbench/common/stunner/bpmn/definition/models/bpsim/FloatingParameter.java
@@ -22,16 +22,16 @@ import javax.xml.bind.annotation.XmlAttribute;
 public class FloatingParameter {
 
     @XmlAttribute
-    private int value;
+    private float value;
 
     // All code behind this comment is auto generated.
     // Please regenerate it again if you added new property.
 
-    public int getValue() {
+    public float getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(float value) {
         this.value = value;
     }
